Refactored rules
