+++ src/NuGetPackageVerifier/PackageIssueFactory.cs
@@ -262,17 +262,17 @@ namespace NuGetPackageVerifier
 
         public static PackageVerifierIssue DependencyVersionHasUpperBound(string id, string dependencyId)
         {
-            return new PackageVerifierIssue("PACKAGE_DEPENDENCY", id, $"The version range of package dependency '{dependencyId}' for package '{id}' has an upper bound.", PackageIssueLevel.Warning);
+            return new PackageVerifierIssue("PACKAGE_DEPENDENCY_VERSION_UPPER_BOUND", id, $"The version range of package dependency '{dependencyId}' for package '{id}' has an upper bound.", PackageIssueLevel.Warning);
         }
 
         public static PackageVerifierIssue DependencyVersionDoesNotHaveLowerBound(string id, string dependencyId)
         {
-            return new PackageVerifierIssue("PACKAGE_DEPENDENCY", id, $"The version range of package dependency '{dependencyId}' for package '{id}' does not have a lower bound.", PackageIssueLevel.Warning);
+            return new PackageVerifierIssue("PACKAGE_DEPENDENCY_VERSION_LOWER_BOUND", id, $"The version range of package dependency '{dependencyId}' for package '{id}' does not have a lower bound.", PackageIssueLevel.Warning);
         }
 
-        public static PackageVerifierIssue DependencyVersionIsPrereleaseForRTMPackage(string id, string dependencyId)
+        public static PackageVerifierIssue DependencyVersionIsPrereleaseForRTMPackage(string id, NuGetVersion version, string dependencyId, NuGetVersion dependencyVersion)
         {
-            return new PackageVerifierIssue("PACKAGE_DEPENDENCY", id, $"The package dependency'{dependencyId}' for package '{id}' is a prerelease package version.", PackageIssueLevel.Error);
+            return new PackageVerifierIssue("PACKAGE_DEPENDENCY_PRERELEASE", id, $"The RTM package '{id}' {version} cannot depend on a pre-release package '{dependencyId}' {dependencyVersion}.", PackageIssueLevel.Error);
         }
     }
 }
